delete the old pidcme and add the new macro to path PWGCF/Flow
