+++ PWGCF/Flow/Tasks/CMakeLists.txt
@@ -48,3 +48,8 @@ o2physics_add_dpl_workflow(flow-gfw-omegaxi
                     SOURCES flowGFWOmegaXi.cxx
                     PUBLIC_LINK_LIBRARIES O2::Framework O2Physics::AnalysisCore O2Physics::GFWCore
                     COMPONENT_NAME Analysis)
+
+o2physics_add_dpl_workflow(flow-pid-cme
+                    SOURCES pidcme.cxx
+                    PUBLIC_LINK_LIBRARIES O2::Framework O2Physics::AnalysisCore O2Physics::GFWCore
+                    COMPONENT_NAME Analysis)
