+++ PWGCF/Tasks/CMakeLists.txt
@@ -28,8 +28,3 @@ o2physics_add_dpl_workflow(correlations
                     SOURCES correlations.cxx
                     PUBLIC_LINK_LIBRARIES O2::Framework O2Physics::AnalysisCore O2Physics::PWGCFCore
                     COMPONENT_NAME Analysis)
-
-o2physics_add_dpl_workflow(pid-cme
-                    SOURCES pidcme.cxx
-                    PUBLIC_LINK_LIBRARIES O2::Framework O2Physics::AnalysisCore O2Physics::PWGCFCore
-                    COMPONENT_NAME Analysis)
